+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -32,8 +32,8 @@ fbf353eb38ae68d8ded3f2a60b432c7bb2c245d2ec7e7c9f53c6da19a0db0938  module/channel
 91425ad75ddb58a3796f7bc88570cb2c81b3db0dd6521a7ce0e55f9148880a47  module/channel-ingress-runtime
 c97dd36cdf8f83c2893c33e9430a93cd131a03d855725783ca5b545de0cf84f8  module/channel-lifecycle
 159d034b431d113f3a6dc41ec0bcadba2d6664051f158330b0e3dd3da8b5d42f  module/channel-logging
-5eccd03e85c032753e0d8931dace4cd952b83e4975d0acae272849fdabfbfe90  module/channel-message
-e2b4d1923a19b927e912622576d43d20a6e188c124a7bc588fe567cd3c1c924e  module/channel-outbound
+40af8b0d826f03a8d0e30213f3ee747a2f54cc8b50934b206d94718ecd10c396  module/channel-message
+df96d17b3a2dd65f8def0be14c98831b49aca96b21f1f08a326f5f7c92ac0cda  module/channel-outbound
 930beff13ed42a138f65164013c82f4f4c96422292c5fc634a5edee1dce71367  module/channel-pairing
 ee4292b069d4d48cce4fc2dc26df5b5c87eb1fa4769f1f6be9a10c3e1221e1a9  module/channel-plugin-common
 94ef57c8f6087fcaa56e59e493c391a04377ed03f23c681edd8d8f6e2d64e0da  module/channel-policy
