+++ src/infra/outbound/deliver-transcript.ts
@@ -1,0 +1,64 @@
+// Mirrors successful outbound payloads into the configured session transcript.
+import { resolveMirroredTranscriptText } from "../../config/sessions/transcript-mirror.js";
+import { createSubsystemLogger } from "../../logging/subsystem.js";
+import { createLazyRuntimeModule } from "../../shared/lazy-runtime.js";
+import { formatErrorMessage } from "../errors.js";
+import type { DeliverOutboundPayloadsCoreParams } from "./deliver-contracts.js";
+import type { NormalizedOutboundPayload } from "./payloads.js";
+import type { OutboundChannel } from "./targets.js";
+
+const log = createSubsystemLogger("outbound/deliver");
+const loadTranscriptRuntime = createLazyRuntimeModule(
+  () => import("../../config/sessions/transcript.runtime.js"),
+);
+
+export async function mirrorDeliveredPayloads(params: {
+  delivery: DeliverOutboundPayloadsCoreParams;
+  payloads: readonly NormalizedOutboundPayload[];
+  channel: Exclude<OutboundChannel, "none">;
+  to: string;
+}): Promise<void> {
+  const mirror = params.delivery.mirror;
+  if (!mirror || params.payloads.length === 0) {
+    return;
+  }
+  const deliveredMirror = {
+    text: params.payloads
+      .map((payload) => payload.hookContent ?? payload.text)
+      .filter((text) => text.trim())
+      .join("\n"),
+    mediaUrls: params.payloads.flatMap((payload) => payload.mediaUrls),
+  };
+  const mirrorText = resolveMirroredTranscriptText({
+    text: deliveredMirror.text,
+    mediaUrls: deliveredMirror.mediaUrls,
+  });
+  if (!mirrorText) {
+    return;
+  }
+  // Transcript mirroring is best-effort bookkeeping after platform send.
+  // Keep mirror failures non-fatal so callers do not retry an already-sent payload.
+  try {
+    const { appendAssistantMessageToSessionTranscript } = await loadTranscriptRuntime();
+    const mirrorResult = await appendAssistantMessageToSessionTranscript({
+      agentId: mirror.agentId,
+      sessionKey: mirror.sessionKey,
+      expectedSessionId: mirror.expectedSessionId,
+      text: mirrorText,
+      idempotencyKey: mirror.idempotencyKey,
+      deliveryMirror: mirror.deliveryMirror,
+      config: params.delivery.cfg,
+    });
+    if (!mirrorResult.ok) {
+      log.warn(
+        `failed to mirror outbound delivery into session transcript; channel send already succeeded: ${mirrorResult.reason}`,
+        { channel: params.channel, to: params.to, sessionKey: mirror.sessionKey },
+      );
+    }
+  } catch (err) {
+    log.warn(
+      `failed to mirror outbound delivery into session transcript; channel send already succeeded: ${formatErrorMessage(err)}`,
+      { channel: params.channel, to: params.to, sessionKey: mirror.sessionKey },
+    );
+  }
+}
